Integrated /jobs with reminder system
