+++ src/pieces/tasks.ts
@@ -1,4 +1,4 @@
-import { BOT, CHANNELS, DB } from '@root/config';
+import { BOT, CHANNELS, DB, MAP_KEY } from '@root/config';
 import { AttachmentBuilder, ChannelType, Client, EmbedBuilder, TextChannel } from 'discord.js';
 import { schedule } from 'node-cron';
 import { Reminder } from '@lib/types/Reminder';
@@ -10,6 +10,8 @@ import { sendToFile } from '../lib/utils/generalUtils';
 import { JobData } from '../lib/types/JobData';
 import { Interest } from '../lib/types/Interest';
 import { JobResult } from '../lib/types/JobResult';
+import { JobPreferences } from '../lib/types/JobPreferences';
+import axios from 'axios';
 
 async function register(bot: Client): Promise<void> {
 	schedule('0/30 * * * * *', () => {
@@ -87,22 +89,22 @@ async function checkPolls(bot: Client): Promise<void> {
 // eslint-disable-next-line no-warning-comments
 async function getJobFormData(userID:string, filterBy: string):Promise<[JobData, Interest, JobResult[]]> {
 	const client = await MongoClient.connect(DB.CONNECTION, { useUnifiedTopology: true });
-	const db = client.db(BOT.NAME).collection(DB.JOB_FORMS);
-	const jobformAnswers:Job[] = await db.find({ owner: userID }).toArray();
-	const jobData:JobData = {
-		city: jobformAnswers[0].answers[0],
-		preference: jobformAnswers[0].answers[1],
-		jobType: jobformAnswers[0].answers[2],
-		distance: jobformAnswers[0].answers[3],
-		filterBy: filterBy ?? 'default'
+	const db = client.db(BOT.NAME).collection(DB.USERS);
+	const jobformAnswers: JobPreferences | null = (await db.findOne({ discordId: userID }))?.jobPreferences;
+	const jobData: JobData = {
+		city: jobformAnswers.answers.city,
+		preference: jobformAnswers.answers.workType,
+		jobType: jobformAnswers.answers.employmentType,
+		distance: jobformAnswers.answers.travelDistance,
+		filterBy: 'default'
 	};
 
-	const interests:Interest = {
-		interest1: jobformAnswers[1].answers[0],
-		interest2: jobformAnswers[1].answers[1],
-		interest3: jobformAnswers[1].answers[2],
-		interest4: jobformAnswers[1].answers[3],
-		interest5: jobformAnswers[1].answers[4]
+	const interests: Interest = {
+		interest1: jobformAnswers.answers.interest1,
+		interest2: jobformAnswers.answers.interest2,
+		interest3: jobformAnswers.answers.interest3,
+		interest4: jobformAnswers.answers.interest4,
+		interest5: jobformAnswers.answers.interest5
 	};
 
 	const APIResponse:JobResult[] = await fetchJobListings(jobData, interests);
@@ -116,14 +118,43 @@ function formatCurrency(currency:number): string {
 	}).format(Number(currency))}`;
 }
 
+function calculateDistance(lat1: number, lon1: number, lat2: number, lon2: number): number {
+	const toRadians = (degrees: number) => degrees * (Math.PI / 180);
+
+	const R = 3958.8; // Radius of the Earth in miles
+	const dLat = toRadians(lat2 - lat1);
+	const dLon = toRadians(lon2 - lon1);
+	const a = Math.sin(dLat / 2) * Math.sin(dLat / 2) +
+		Math.cos(toRadians(lat1)) * Math.cos(toRadians(lat2)) *
+		Math.sin(dLon / 2) * Math.sin(dLon / 2);
+	const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
+	const distance = (lat1 === 0 && lon1 === 0) || (lat2 === 0 && lon2 === 0) ? -1 : R * c;
+	return distance;
+}
+
+async function queryCoordinates(location: string): Promise<{lat: number, lng: number}> {
+	const preferredCity = encodeURIComponent(location);
+
+	const baseURL = `https://maps.google.com/maps/api/geocode/json?address=${preferredCity}&components=country:US&key=${MAP_KEY}`;
+	const response = await axios.get(baseURL);
+	const coordinates : {lat: number, lng: number } = {
+		lat: response.data.results[0].geometry.location.lat,
+		lng: response.data.results[0].geometry.location.lng
+	};
+
+	return coordinates;
+}
+
 function titleCase(jobTitle:string): string {
 	return jobTitle.toLowerCase().replace(/[()]/g, '').split(' ').map(word => word.charAt(0).toUpperCase() + word.slice(1)).join(' ');
 }
 
-function listJobs(jobData: JobResult[], filterBy: string): string {
+async function listJobs(jobForm: [JobData, Interest, JobResult[]], filterBy: string): Promise<string> {
 	// Conditionally sort jobs by salary if sortBy is 'salary'
+	const cityCoordinates = await queryCoordinates(jobForm[0].city);
+
 	if (filterBy === 'salary') {
-		jobData.sort((a, b) => {
+		jobForm[2].sort((a, b) => {
 			const avgA = (Number(a.salaryMax) + Number(a.salaryMin)) / 2;
 			const avgB = (Number(b.salaryMax) + Number(b.salaryMin)) / 2;
 
@@ -133,40 +164,64 @@ function listJobs(jobData: JobResult[], filterBy: string): string {
 
 			return avgB - avgA; // Descending order
 		});
+	} else if (filterBy === 'alphabetical') {
+		jobForm[2].sort((a, b) => (a.title > b.title ? 1 : -1));
+	} else if (filterBy === 'date') {
+		jobForm[2].sort((a, b) => new Date(b.created).getTime() - new Date(a.created).getTime());
+	} else if (filterBy === 'distance') {
+		// cityCoordinates = await this.queryCoordinates(jobForm[0].city);
+
+		jobForm[2].sort((a, b) => {
+			const distanceA = calculateDistance(cityCoordinates.lat, cityCoordinates.lng, Number(a.latitude), Number(a.longitude));
+			const distanceB = calculateDistance(cityCoordinates.lat, cityCoordinates.lng, Number(b.latitude), Number(b.longitude));
+
+			if (distanceA === -1) {
+				return 1; // Treat jobs with no location as lowest
+			}
+
+			return distanceA - distanceB; // Might have to account for negative distances
+		});
 	}
 
+
+
 	let jobList = '';
-	for (let i = 0; i < jobData.length; i++) {
-		const avgSalary = (Number(jobData[i].salaryMax) + Number(jobData[i].salaryMin)) / 2;
+	for (let i = 0; i < jobForm[2].length; i++) {
+		const avgSalary = (Number(jobForm[2][i].salaryMax) + Number(jobForm[2][i].salaryMin)) / 2;
 		const formattedAvgSalary = formatCurrency(avgSalary);
-		const formattedSalaryMax = formatCurrency(Number(jobData[i].salaryMax)) !== 'N/A' ? formatCurrency(Number(jobData[i].salaryMax)) : '';
-		const formattedSalaryMin = formatCurrency(Number(jobData[i].salaryMin)) !== 'N/A' ? formatCurrency(Number(jobData[i].salaryMin)) : '';
+		const formattedSalaryMax = formatCurrency(Number(jobForm[2][i].salaryMax)) !== 'N/A' ? formatCurrency(Number(jobForm[2][i].salaryMax)) : '';
+		const formattedSalaryMin = formatCurrency(Number(jobForm[2][i].salaryMin)) !== 'N/A' ? formatCurrency(Number(jobForm[2][i].salaryMin)) : '';
+		const jobDistance = calculateDistance(cityCoordinates.lat, cityCoordinates.lng, Number(jobForm[2][i].latitude), Number(jobForm[2][i].longitude));
+		const formattedDistance = (jobDistance !== -1) ? `${jobDistance.toFixed(2)} miles` : 'N/A';
 
 		const salaryDetails = (formattedSalaryMin && formattedSalaryMax)
 			? `, Min: ${formattedSalaryMin}, Max: ${formattedSalaryMax}`
 			: formattedAvgSalary;
 
-		jobList += `${i + 1}. **${titleCase(jobData[i].title)}**  
+		jobList += `${i + 1}. **${jobForm[2][i].title}**  
 		  \t\t* **Salary Average:** ${formattedAvgSalary}${salaryDetails}  
-		  \t\t* **Location:** ${jobData[i].location}  
-		  \t\t* **Apply here:** [read more about the job and apply here](${jobData[i].link})  
-		  ${i !== jobData.length - 1 ? '\n' : ''}`;
+		  \t\t* **Location:** ${jobForm[2][i].location}  
+		  \t\t* **Date Posted:** ${new Date(jobForm[2][i].created).toDateString()} at ${new Date(jobForm[2][i].created).toLocaleTimeString('en-US', { hour: '2-digit', minute: '2-digit' })}
+		  \t\t* **Apply here:** [read more about the job and apply here](${jobForm[2][i].link})  
+		  \t\t* **Distance:** ${formattedDistance}
+		  ${i !== jobForm[2].length - 1 ? '\n' : ''}`;
 	}
 
 	return jobList || '### Unfortunately, there were no jobs found based on your interests :(. Consider updating your interests or waiting until something is found.';
 }
 
-async function jobMessage(reminder: Reminder, userID: string): Promise<string> {
-	const jobFormData: [JobData, Interest, JobResult[]] = await getJobFormData(userID, reminder.filterBy);
-	const message = `## Hey <@${reminder.owner}>!  
+export async function jobMessage(reminder: Reminder | string, userID: string): Promise<string> {
+	const jobFormData: [JobData, Interest, JobResult[]] = await getJobFormData(userID, typeof reminder === 'object' && 'filterBy' in reminder ? reminder.filterBy : 'default');
+	const filterBy = typeof reminder === 'object' && 'filterBy' in reminder ? String((reminder as Reminder).filterBy) : String(reminder);
+	const message = `## Hey <@${userID}>!  
 	## Here's your list of job/internship recommendations:  
-	Based on your interests in **${jobFormData[1].interest1}**, **${jobFormData[1].interest2}**, \
-	**${jobFormData[1].interest3}**, **${jobFormData[1].interest4}**, and **${jobFormData[1].interest5}**, I've found these jobs you may find interesting. Please note that while you may get\
-	job/internship recommendations from the same company,\
-	their positions/details/applications/salary WILL be different and this is not a glitch/bug!
-	Here's your personalized list:
+Based on your interests in **${jobFormData[1].interest1}**, **${jobFormData[1].interest2}**, \
+**${jobFormData[1].interest3}**, **${jobFormData[1].interest4}**, and **${jobFormData[1].interest5}**, I've found these jobs you may find interesting. Please note that while you may get\
+job/internship recommendations from the same company,\
+their positions/details/applications/salary WILL be different and this is not a glitch/bug!
+Here's your personalized list:
 
-	${listJobs(jobFormData[2], reminder.filterBy)}
+	${await listJobs(jobFormData, filterBy)}
 	---  
 	### **Disclaimer:**  
 	-# Please be aware that the job listings displayed are retrieved from a third-party API. \
@@ -178,7 +233,7 @@ async function jobMessage(reminder: Reminder, userID: string): Promise<string> {
 	return message;
 }
 
-function stripMarkdown(message: string, owner: string): string {
+export function stripMarkdown(message: string, owner: string): string {
 	return message
 		.replace(new RegExp(`## Hey <@${owner}>!\\s*## Here's your list of job/internship recommendations:?`, 'g'), '') // Remove specific header
 		.replace(/\[read more about the job and apply here\]/g, '')
@@ -190,7 +245,7 @@ function stripMarkdown(message: string, owner: string): string {
 		.trim();
 }
 
-function headerMessage(owner:string, filterBy:string):string {
+export function headerMessage(owner:string, filterBy:string):string {
 	return `## Hey <@${owner}>!  
 	### **__Please read this disclaimer before reading your list of jobs/internships__:**  
 -# Please be aware that the job listings displayed are retrieved from a third-party API. \
